requirements.txt
autopep8==1.5.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==2.2
django-debug-toolbar==3.2.1
django-debug-toolbar-template-profiler==2.0.2
django-environ==0.4.5
idna==2.10
oauthlib==3.1.0
Pillow==8.1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
